requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.3.0
langgraph>=0.2.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.0.0

# File Parsing (fallback)
pdfplumber>=0.10.0
python-docx>=1.0.0

# Web App
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
